// src/api/generated/model/createApplicationGraderRequest.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export interface CreateApplicationGraderRequest {
  /** @minLength 1 */
  userId: string;
  /** @minLength 1 */
  applicationId: string;
  endEndDate?: string;
  orderNumber: number;
  isReferee?: boolean;
}