// src/api/generated/model/updateApplicationGraderRequest.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateApplicationGraderRequest {
  endEndDate?: string;
  isCompleted?: boolean;
  orderNumber?: number;
  isReferee?: boolean;
}